fix: move new endpoints to api
