+++ oioioi/problems/views.py
@@ -26,7 +26,7 @@ from django.db.models import (
     When,
 )
 from django.db.models.functions import Coalesce
-from django.http import Http404, HttpResponse, HttpResponseBadRequest, JsonResponse
+from django.http import Http404, HttpResponse, HttpResponseBadRequest
 from django.shortcuts import get_object_or_404, redirect
 from django.template.response import TemplateResponse
 from django.urls import reverse
@@ -34,7 +34,7 @@ from django.utils.encoding import force_str
 from django.utils.safestring import mark_safe
 from django.utils.translation import get_language
 from django.utils.translation import gettext_lazy as _
-from django.views.decorators.http import require_http_methods, require_safe
+from django.views.decorators.http import require_http_methods
 from unidecode import unidecode
 
 from oioioi.base.permissions import enforce_condition, not_anonymous
@@ -92,7 +92,7 @@ from oioioi.problems.utils import (
     query_statement,
     show_proposal_form,
 )
-from oioioi.programs.models import ModelSolution, Test
+from oioioi.programs.models import ModelSolution
 
 
 def show_statement_view(request, statement_id):
@@ -567,68 +567,6 @@ def problem_site_external_attachment_view(request, site_key, attachment_id):
     if attachment.problem.id != problem.id:
         raise PermissionDenied
     return stream_file(attachment.content, attachment.download_name)
-
-
-@require_safe
-def problem_site_example_tests_view(request, site_key):
-    problem = get_object_or_404(Problem, problemsite__url_key=site_key)
-    tests = Test.objects.filter(
-        problem_instance=problem.main_problem_instance,
-        kind="EXAMPLE",
-    ).order_by("order", "name")
-
-    result = []
-    for test in tests:
-        basename = f"{problem.short_name}{test.name}"
-        result.append(
-            {
-                "name": test.name,
-                "in_url": reverse(
-                    "problem_site_example_test_file",
-                    kwargs={"site_key": site_key, "filename": f"{basename}.in"},
-                )
-                if test.input_file
-                else None,
-                "out_url": reverse(
-                    "problem_site_example_test_file",
-                    kwargs={"site_key": site_key, "filename": f"{basename}.out"},
-                )
-                if test.output_file
-                else None,
-            }
-        )
-
-    return JsonResponse(result, safe=False)
-
-
-@require_safe
-def problem_site_example_test_file_view(request, site_key, filename):
-    problem = get_object_or_404(Problem, problemsite__url_key=site_key)
-
-    base, ext = os.path.splitext(filename)
-    if ext not in (".in", ".out"):
-        raise Http404
-
-    short_name = problem.short_name
-    if not base.startswith(short_name):
-        raise Http404
-    test_name = base[len(short_name) :]
-
-    test = get_object_or_404(
-        Test,
-        problem_instance=problem.main_problem_instance,
-        kind="EXAMPLE",
-        name=test_name,
-    )
-
-    if ext == ".in":
-        if not test.input_file:
-            raise Http404
-        return stream_file(test.input_file, filename)
-    else:
-        if not test.output_file:
-            raise Http404
-        return stream_file(test.output_file, filename)
 
 
 def problemset_add_to_contest_view(request, site_key):
